docs: correct the logout server function's authentication comment
logoutFn has been exempt from authentication since before this branch —
exceptions.ts says why — but the comment above it claimed the opposite.
State the real contract: logout is unauthenticated and idempotent, so a
caller whose session already expired still gets its stale cookies cleared
instead of a 401 that would strand the browser holding dead credentials.

Also lift NO_PERMISSIONS and seedGroup into groups/test/fixtures.ts, so
the permission shape is written once rather than in each test that needs
a group.
